moved icons to shared
