+++ apps/gecko/src/popup/components/Navbar/Navbar.tsx
@@ -1,7 +1,7 @@
 import styled from "styled-components";
 import { browser } from "webextension-polyfill-ts";
 import { FILTER_LIST_NAV, PREFERENCE_LIST_NAV } from "../../constants";
-import { CogIcon } from "../../icons";
+import { CogIcon } from "@ui/icons";
 import NavItem from "./NavItem";
 
 interface Props {
