+++ libs/common/ui/icons/CogIcon.tsx
