+++ libs/common/ui/icons/DragHandleIcon.tsx
