+++ libs/common/ui/icons/PencilIcon.tsx
