+++ libs/common/ui/icons/TrashIcon.tsx
